Add information about running an OpenMM simulation
Added some information about how to run an OpenMM simulation from the command line and linked to the example Jupyter notebook for creating and serialising an OpenMM system, and running this in NanoVer.

I've also added a link to the documentation page for the OpenMM serialiser, however, this page is currently empty. This will be fixed with this issue on the Nanover documentation repo: https://github.com/IRL2/nanover/issues/10